Changed AzureFunctions from In-Process mode to isolated worker process.
