+++ src/TransCelerate.SDR.sln
@@ -15,7 +15,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransCelerate.SDR.Services"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransCelerate.SDR.RuleEngine", "TransCelerate.SDR.RuleEngine\TransCelerate.SDR.RuleEngine.csproj", "{FC722F3A-20A4-4C0A-BEE2-DC87C736D593}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TransCelerate.SDR.AzureFunctions", "TransCelerate.SDR.AzureFunctions\TransCelerate.SDR.AzureFunctions.csproj", "{38F6018E-2D4E-4E49-9287-2D656AC9DC70}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransCelerate.SDR.AzureFunctions", "TransCelerate.SDR.AzureFunctions\TransCelerate.SDR.AzureFunctions.csproj", "{F5C69AF1-320D-41E6-AA7F-EAB923B49B05}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -47,10 +47,10 @@ Global
 		{FC722F3A-20A4-4C0A-BEE2-DC87C736D593}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FC722F3A-20A4-4C0A-BEE2-DC87C736D593}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FC722F3A-20A4-4C0A-BEE2-DC87C736D593}.Release|Any CPU.Build.0 = Release|Any CPU
-		{38F6018E-2D4E-4E49-9287-2D656AC9DC70}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{38F6018E-2D4E-4E49-9287-2D656AC9DC70}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{38F6018E-2D4E-4E49-9287-2D656AC9DC70}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{38F6018E-2D4E-4E49-9287-2D656AC9DC70}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F5C69AF1-320D-41E6-AA7F-EAB923B49B05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F5C69AF1-320D-41E6-AA7F-EAB923B49B05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F5C69AF1-320D-41E6-AA7F-EAB923B49B05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F5C69AF1-320D-41E6-AA7F-EAB923B49B05}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
